Release 1.0.2 * Implemented plugin wrapping layer for more flexible plugin control. * Added GUI wrapper for LinuxVST plugins.

+++ include/plugins/phase_detector.h
@@ -59,7 +59,7 @@ namespace lsp
         public:
             virtual void destroy();
             virtual void update_settings();
-            virtual void update_sample_rate(int sr);
+            virtual void update_sample_rate(long sr);
             virtual void process(size_t samples);
     };
